requirements-dev.txt
# Development dependencies for pulse

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code quality
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
isort>=5.12.0

# Type stubs
types-PyYAML>=6.0.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.0.0

# Build
build>=0.10.0
twine>=4.0.0

# Optional runtime dependencies for testing
PyYAML>=6.0
dnspython>=2.3.0
